Normalize source indentation
The AssertJ source blocks are taken from real code, which is indented as
usual in an IDE. To avoid the source snippets use that (somewhat random)
indentation in the documentation afterwards, the indentation of each
source block needs to be set to zero.
While that is currently done on each block separately, it is less
error-prone by using the globally available attribute.